Few big changes here:
- added support for 'secret' - as a normal request attributes. If
the handler is configured to require a secret, it'll block any request
without one. If not - older mod_jk's will work. ( I'll commit the change
on the C code for both mod_jk and mod_jk2 )

- added support for 'ajp13.id' - similar with what ajp12interceptor is
doing in 3.3. This will save the connection info ( including secret )
for easy reading by caller. ( not completed )

- add useSecret option - it'll automatically generate a secret, no
config needed ( and no need to ask the user to change the pass
regularily - we'll do it on each server restart ). Mod_jk will just
read the file.

- added support for the shutdown message ( not completed )

The request handler will call 'container' as the next handler,
unless overriden. It'll also check if a dispathcer is
present and register the messages it supports ( right now it's
the only handler, so dispatcher is optional )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296369 13f79535-47bb-0310-9956-ffa450edef68