Implementation of the 'secret' attribute in the request, used to pass a key used to authenticate the sender.
This shouldn't affect in any way the current code - as long as the
secret attribute is not set on the worker, the code will behave
exactly as before and work with any tomcat from 3.2 to 4.1.

If the attribute is set, the attribute will be sent - the
receiving side should know how to deal with it, that means
it should use a recent version of jk.

On the java side, the new versions of jk will allow setting the
key - that will require any web server to send the password
in order to allow forwarding the request. The check will happen
once per tcp connection.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296379 13f79535-47bb-0310-9956-ffa450edef68