Made some of the requtil functions part of the jk_service interface.
Server adapters can override those methods - for example a more
efficient implementation of getMethod() is possible, serializing
the data directly on the jk_msg, without copy, etc.

In addition this allows server adapters to integrate with other
server modules - for example the cookie ( required for lb ) may
be already parsed, or the session id may be integrated with
other resources ( this will be required as soon as we implment
form auth ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296587 13f79535-47bb-0310-9956-ffa450edef68