Remove the old jni code, no longer supported for jk2.
Jk2 uses the a jni channel and ajp13 marshalling. This minimize the amount
of JNI calls ( each cross has a performance impact ) and eliminate the allocation
and conversion on the C side ( creating the strings was hugely expensive,
almost eliminating any benefit of doing jni - the java side knows how to
avoid strings and create them efficiently, and the optimizations is hard
to reproduce in C ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296629 13f79535-47bb-0310-9956-ffa450edef68