Remove WEB-INF.
Added a flag to build APR-based jk with apache13
It should work, but the build file is broken, it includes
apr by using the apache2 include dir - if detected. That creates
problem since httpd.h is there too.

We need a standalone-APR to use with apache13 ( and IIS, etc - as we add
them ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296740 13f79535-47bb-0310-9956-ffa450edef68