Add jk_shm to the registry.
The general structure of the shared memory will be very simple - a number
 of 'slots', each following its own rules. That will allow multiple
components to use the shm.

My goal is to get at least 2 'slots' working - one for the worker status,
similar with mod_jserv, and one for the config 'generation', allowing
apache processes to refresh themself when config changes.
( this is not the most urgent piece for jk2, but I wanted to get it at least
in a basic state - maybe other people will help )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296755 13f79535-47bb-0310-9956-ffa450edef68