Added a pointer to the shm object.
Note that the 'shm' doesn't depend directly on APR or shared memory availability.
Other implementations are possible to support 'strange' OSes or
workarounds, or 'native' libraries.

For what we need it may not even be necesary to use shared memory -
another form of IPC could also work as long as it allow some way to
keep some data in (relative) sync among processes.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296757 13f79535-47bb-0310-9956-ffa450edef68