A simple component that will automatically register the tomcat workers in the jk 'scoreboard'.
Getting this to work is the last step in supporting multi-process JNI and
dynamic addition/removal/status of workers for jk.

So far attach() seems to work ( and I'm not sure if tomcat should create the
shmem ).

The data structures in the shm will be very similar with what jserv used to have,
but more flexible.

We may also need to add some mutexes - or at least atomic operations from
APR.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296845 13f79535-47bb-0310-9956-ffa450edef68