Allow a 'range' of ports to be used.
This allows multiple tomcats to be started on the same machine with the
same configuration - and get consecutive sockets ( tcp and unix ) automatically.
Needed for jni, it also simplify the configuration of lb workers - no need
to use separate config files for each worker ( on a machine ).

The port will also determine the local part of the instance ID ( the hostname
or address should form the other part - and we need a way to push this up
into the coyote and container ).

If Shm is detected it'll set the instance id based on the shared memory
protocol ( i.e. each worker will get a unique id based on the position
of the worker def in the shared mem ) .


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296870 13f79535-47bb-0310-9956-ffa450edef68