Few more features. I think the sending part works fine now ( but we should still do more tunning ).
For each tomcat instance we create a shm slot in the scoreboard. The
slot is named TOMCAT:host:port ( host:port is used as instance ID ), and
contains a ( ajp13-encoded ) structure with informations about the channels
( currently the socket and unix channel - jni is special ). It'll also
include the groups where the instance belongs, and possibly other
informations.

If the number of channels is 0, the lb worker is supposed to stop
sending new requests to the instance ( but keep forwarding the
existing sessions ). That'll implement 'gracefull shutdown'.

Still to add:
- display ( nicely ) the information in the shm
- read info from shm ( like slots owned by apache processes with
statistics - in a future release )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297026 13f79535-47bb-0310-9956-ffa450edef68