That's the big one.
Please review !

It changes the handling of lb_value to int. I also cleaned up the logic so
it's easier ( I hope ) to understand what's happening. "Levels" replace
the 'local worker', I thing I got the logic straight for those.

I started to add a 'introspection' data, to validate and better report
the config.

We use one table per level. At the moment the maximum number of workers
is hardcoded ( to 255 ), we could make it dynamic but that would make things
pretty complex when we add workers dynamically ( it won't work without
a CS or atomic operations )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297123 13f79535-47bb-0310-9956-ffa450edef68